Extract video player event subscription to hook and added more video status
